Extract most of the version/changelog check into a validator
